+++ src/LanguageJournal/Controllers/UsersController.cs
@@ -13,44 +13,60 @@ namespace LanguageJournal.Controllers {
     [Route("api/[controller]")]
     public class UsersController : Controller {
         private readonly PostgresDbContext _db;
+        private Authenticator _auth;
 
         public UsersController(PostgresDbContext context) {
             _db = context;
+            _auth = new Authenticator(_db);
         }
 
         // GET api/users
         [HttpGet]
         public IEnumerable<User> Get() {
-            return _db.Users.ToList();
+            return _auth.AuthenticateWithResult<IEnumerable<User>>(HttpContext, _db.Users.ToList());
         }
 
         // GET api/users/5
         [HttpGet("{id}")]
         public User Get(int id) {
-            return _db.Users.Where(u => u.UserId == id).FirstOrDefault();
+            return _auth.AuthenticateWithResult<User>(HttpContext, _db.Users.Where(u => u.UserId == id).FirstOrDefault());
         }
 
         // POST api/users
         [HttpPost]
         public User Post([FromBody]User user) {
-            _db.Users.Add(user);
-            _db.SaveChanges();
-            return user;
+            var authUser = _auth.AuthenticateHeaders(HttpContext);
+            if (authUser != null) {
+                _db.Users.Add(user);
+                _db.SaveChanges();
+                return user;
+            } else {
+                return null;
+            }
 
         }
 
         // PUT api/users/5
         [HttpPut("{id}")]
         public User Put(int id, [FromBody]User user) {
-            user.UserId = id;
-            _db.Entry(user).State = EntityState.Modified;
-            _db.SaveChanges();
-            return user;
+            var authUser = _auth.AuthenticateHeaders(HttpContext);
+            if (authUser != null) {
+                user.UserId = id;
+                _db.Entry(user).State = EntityState.Modified;
+                _db.SaveChanges();
+                return user;
+            } else {
+                return null;
+            }
         }
 
         // DELETE api/users/5
         [HttpDelete("{id}")]
         public void Delete(int id) {
+            var authUser = _auth.AuthenticateHeaders(HttpContext);
+            if (authUser == null) {
+                return;
+            }
             var user = _db.Users.Where(u => u.UserId == id).FirstOrDefault();
             _db.Users.Remove(user);
             _db.SaveChanges();
